Refactor welcome panel and add UI scale binding
