--- Ciallo/GuiControl/RecentDocumentList.cs
@@ -1,25 +1,0 @@
-using Godot;
-
-namespace Ciallo.GuiControl;
-
-public partial class RecentDocumentList : ItemList
-{
-    public override void _Ready()
-    {
-        this.ObserveObservableList(AppPreference.RecentFiles);
-        ItemActivated += idx =>
-        {
-            var path = AppPreference.RecentFiles[(int)idx];
-            var success = OpenDocumentDialog.LoadWorldFile(path);
-            if (!success) AppPreference.RecentFiles.Remove(path);
-        };
-
-        ItemClicked += (index, _, buttonIndex) =>
-        {
-            if ((MouseButton)buttonIndex == MouseButton.Right)
-            {
-                AppPreference.RecentFiles.RemoveAt((int)index);
-            }
-        };
-    }
-}
